fix: avoid sharing state between components with same props
Before we memoized the component closures based on the props in order to
keep the re-renders consistent. It led to an issue in which components
with the same props shared state.
Now it is solved by creating an id for every component definition call,
giving different closure refs but keeping consistent between renders.